+++ test/libsolidity/smtCheckerTests/special/msg_sender_2.sol
@@ -10,5 +10,3 @@ contract C
 	}
 }
 // ----
-// Warning: (98-108): Assertion checker does not yet implement this expression.
-// Warning: (98-108): Internal error: Expression undefined for SMT solver.